・moduleのDecoratorを使ってviewに関わるメソッドを定義 ・タブをページごとに切り替え
