+++ app/controllers/application_controller.rb
@@ -1,2 +1,6 @@
 class ApplicationController < ActionController::Base
+    def current_user
+      ActiveDecorator::Decorator.instance.decorate(super) if super.present?
+      super
+    end
 end
