+++ app/models/user.rb
@@ -44,10 +44,6 @@ class User < ApplicationRecord
     likes.exists?(article_id: article.id)
   end
 
-  def display_name
-    profile&.nickname || self.email.split('@').first
-  end
-
   def follow!(user)
     user_id = get_user_id(user)
     following_relationships.create!(following_id: user_id)
@@ -69,13 +65,6 @@ class User < ApplicationRecord
     profile || build_profile
   end
 
-  def avatar_image
-     if profile&.avatar&.attached?
-      profile.avatar
-     else
-      'default-avatar.png'
-      end
-   end
 
    private 
    def get_user_id(user)
